report/requirements.txt
redshift_connector
python-dotenv
jinja2
pytest